requirements.txt
fastapi[standard]
agno
google-genai
groq
pydantic
pydantic-settings
qdrant-client
sentence-transformers
sqlmodel
openai
jwt
PyJWT 
pymupdf
pymupdf-stubs
# --- Security ---
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
qdrant-client[fastembed]>=1.10.0